deprecated shape_dfs
